requirements.txt
aiogram==2.25.1
beautifulsoup4==4.12.2
pdfkit==1.0.0
python-dotenv==1.0.0
